First commit.
I added a feature for watch a list ToDo, is a example from the book The little book of ASP.NET Core.

I created classes in the folder Models, Views and created the folder Services for the app s services with two classes.

And from now i will writer in english all comments, codes and commits with few help of Traslator.